babymind-os/requirements.txt
# BabyMind OS - 육아 AI 인텔리전스 CCTV MCP 서비스
# ===================================================

# AI 분석
anthropic>=0.40.0          # Claude Vision API (프레임 분석)

# 영상 스트림 처리
opencv-python-headless>=4.9.0  # RTSP 스트림 캡처 (headless: 서버용)
Pillow>=10.0.0             # 이미지 처리

# MCP 서버
mcp>=1.0.0                 # Model Context Protocol SDK

# 알림 시스템
resend>=2.0.0              # 이메일 발송 (Resend API)
httpx>=0.27.0              # 비동기 HTTP (카카오톡 API 등)

# 데이터 저장
supabase>=2.0.0            # Supabase (분석 결과 저장)

# 유틸리티
python-dotenv>=1.0.0       # 환경변수
pydantic>=2.0.0            # 데이터 모델 검증
apscheduler>=3.10.0        # 스케줄링 (주기적 분석)
aiohttp>=3.9.0             # 비동기 웹 요청
numpy>=1.24.0              # 수치 계산
